Build style guide

+++ packages/components/utility/build-copy.js
@@ -6,9 +6,13 @@ const getFilesFolders = require('./utility').getFilesFolders
 const buildDir = path.join(__dirname, '../build')
 const deployDir = path.join(__dirname, '../deploy/client')
 
-getFilesFolders(deployDir, true, 'file').forEach(file => {
-  fs.unlinkSync(file)
-})
+if (fs.existsSync(deployDir)) {
+  getFilesFolders(deployDir, true, 'file').forEach(file => {
+    fs.unlinkSync(file)
+  })
+} else {
+  fs.mkdirSync(deployDir)
+}
 
 getFilesFolders(buildDir, true, 'both').forEach(fileFolder => {
   const newLocation = fileFolder.replace(buildDir, deployDir)
